refactor(orientdb): use rest-api instead of client library (#80)
The client library seems not well maintained and contains pretty
outdated dependencies. For testing purposes using the REST-API works
well enough.

`retry` is used to stabilise the test, since at least locally it took
the container some time after start-up to accept REST-API requests
sometimes. I couldn't find out a way to wait exactly for the time when
the API is ready. According to the container logs the start-up is
finished.

Closes #38