feat: add oracle image impl (#140)
Hopefully everything is ok, but if something could be better, I'll be
happy to improve it.

Note that the unit test needs `Oracle client` to run. I'm not sure how
to handle this in the CI pipeline. Unfortunately my experience with
GitHub Actions is very limited, so I'll need some tips.