feat: add Foundry Anvil image impl (#272)
# Community Testcontainers Implementation for [Foundry
Anvil](https://book.getfoundry.sh/anvil/)

This is a community implementation of the
[Testcontainers](https://testcontainers.org/) interface for
[Foundry](https://github.com/foundry-rs/foundry)
[Anvil](https://book.getfoundry.sh/anvil/).

It is not officially supported by Foundry, but it is a community effort
to provide a more user-friendly interface for running Anvil inside a
Docker container.

The endpoint of the container is intended to be injected into your
provider configuration, so that you can easily run tests against a local
Anvil instance. See the `test_anvil_node_container` test in
`src/anvil/mod.rs` for an example.