test(gitea): fix broken test in CI
The test was broken in CI, but may work locally (on MacOS for example).
In CI it fails with: `self-signed certificate`

It's because of different behavior in `rust-native-tls` depending on the platform.
This PR eliminates the issue via explicit usage of rustls.
Alternative would be to disable cert verification, but it means test would be useless.