feat(minio)!: support newest MinIO versions (#303)
It seems like at some point in the last 3 years MinIO switched from
writing startup messages on stdou to stderr. I couldn't find a
`WaitFor::either()` to keep the old version as the default, so I bumped
it to the latest released version.

This update is necessary for folks to be able to use a modern or
different tag with `ImageExt`.

Co-authored-by: Artem Medvedev <i@ddtkey.com>